Move HomeViewModel to correct folder
